+++ MapboxNavigation/RouteMapViewController.swift
@@ -281,6 +281,7 @@ class RouteMapViewController: UIViewController {
     override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
         super.traitCollectionDidChange(previousTraitCollection)
         mapView.enableFrameByFrameCourseViewTracking(for: 3)
+        navigationView.reinstallRequiredConstraints()
     }
 
     override func viewDidLayoutSubviews() {